+++ lib/types/ContextMenu.d.ts
@@ -9,11 +9,11 @@ export interface MenuWrapper<T> {
 }
 export default class ContextMenu {
     private storeMenus;
-    clickEventFunc: (e: MouseEvent) => void;
     contextMenuOption: ContextMenuOption;
     constructor(option?: ContextMenuOption);
     private injectCss;
     private hideMenuEventListener;
+    private clickEventFunc;
     private scrollListener;
     create<Payload>(menuOption: MenuOption<Payload>): MenuWrapper<Payload>;
     showMenu<T>(position: PanelPosition, menu: Menu<T>, payload?: T): void;